Move ImageReference to make it easier to use
