+++ constellation/util.py
@@ -1,0 +1,8 @@
+class ImageReference:
+    def __init__(self, repo, name, tag):
+        self.repo = repo
+        self.name = name
+        self.tag = tag
+
+    def __str__(self):
+        return "{}/{}:{}".format(self.repo, self.name, self.tag)
